Time: 51 ms (60.76%), Space: 19 MB (11.65%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -9,7 +9,7 @@ public:
             
             int sum = 0;
             for (auto &i : piles) {
-                sum += ceil((double)i / k);
+                sum += (int)ceil((double)i / k);
                 if (sum > h) {
                     left = k + 1;
                     continue;
